Migrate configTools from HToZZ4L to RootTools, add support for negative weights
